Minor fix for older return type detection on older compilers

+++ src/tam.c
@@ -55,7 +55,7 @@ size_t first_line(prototype_list_t* list) {
   prototype_node_t* node = NULL;
   for (node = list->first; NULL != node; node = node->next) {
     if (node->info.is_function_implementation) {
-      return node->info.is_function_implementation;
+      return (size_t)node->info.is_function_implementation;
     }
   }
   return 0;
